+++ CHANGELOG.md
@@ -16,6 +16,7 @@ please see [changelog_updates.md](docs/dev/changelog_updates.md).
 #### Minor
 
 - Uptime Kuma is no longer mandatory and the status dashboard can be disabled
+- Added Owning Organization in Admin Connector Overview ([#355](https://github.com/sovity/authority-portal/issues/355))
 
 #### Patch